// src/app/core/config/roles.enum.ts
export enum AppRole {
    SEC = 'SEC',
    Secretary = 'SEC',
    HOD = 'HOD',
    DCL = 'DCL',
    DeputyCommissioner = 'DCL',
    DDO = 'DDO',
    ADMN = 'ADMN',
    Administrator = 'ADMN',
    NDOF = 'NDOF',
    NodalOfficer = 'NDOF',
    PPOF = 'PPOF',
    PetrolPumpOfficer = 'PPOF',
    ROFC = 'ROFC',
    RevenueOfficer = 'ROFC',
    GUEST = 'GUEST',
    Guest = 'GUEST',
    FD = 'FD',
}